+++ fut/config.py
@@ -1,6 +1,6 @@
 # chrome 53 @ win10
 headers = {
-    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/54.0.2840.99 Safari/537.36',
+    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.95 Safari/537.36',
     'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
     'Accept-Encoding': 'gzip,deflate,sdch',
     'Accept-Language': 'en-US,en;q=0.8',
@@ -40,3 +40,4 @@ flash_agent = 'ShockwaveFlash/24.0.0.186'
 
 cookies_file = 'cookies.txt'
 timeout = 15  # defaulf global timeout
+delay = (0, 5)  # default mininum delay between requests (random range)